requirements.txt
faiss-cpu==1.7.2
imageio==2.22.0
glob2==0.7
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.6.0.66
Pillow==9.2.0
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@0d53acbcf5123d4c88040fe68fbb9805fc5b2fb9
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@6ae6ff9cf73221ce60617ef4658b4892b986ba9d
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.7.3
torch==1.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.1